Fixed octals

+++ main.h
@@ -6,12 +6,13 @@
 #include <math.h>
 #include <stdint.h>
 #include <string.h>
+#include <stdio.h>
 
 int _putchar(char);
 int _printf(const char *format, ...);
 int64_t print_number(int64_t n);
-int dec2bin(unsigned int n);
-int dec2oct(unsigned int n);
+int64_t dec2bin(unsigned int n);
+int64_t dec2oct(unsigned int n);
 char *dec2hex(unsigned int n, char _case);
 
 #endif
